QtMobility Memory Corruption Reproduction:
When running DC as root, a "operation cannot be performed" notification
shows up and the contacts tab will display nothing.  If you closed on
that tab, then startup will hault.

When using this script, one can get various issues like floating point
exceptions.

The best I can figure is a bug in QtMobility or PySide that is causing
memory corruption, taking out large swathes of stuff with it